# Improve performance of anomaEncode / anomaDecode in the Core evaluator