--- a/bench2/Benchmark/Nockma.hs
+++ b/bench2/Benchmark/Nockma.hs
@@ -1,11 +1,13 @@
 module Benchmark.Nockma where
 
 import Benchmark.Nockma.Encoding qualified as NockmaEncoding
+import Benchmark.Nockma.Encoding.ByteString qualified as NockmaByteString
 import Test.Tasty.Bench
 
 bm :: Benchmark
 bm =
   bgroup
     "Nockma"
-    [ NockmaEncoding.bm
+    [ NockmaEncoding.bm,
+      NockmaByteString.bm
     ]