--- a/bench2/Benchmark/Nockma/Encoding/ByteString.hs
+++ b/bench2/Benchmark/Nockma/Encoding/ByteString.hs
@@ -10,10 +10,17 @@ randomBytes numBytes = do
   sg <- getStdGen
   return (fst (genByteString numBytes sg))
 
+testBytesSize :: Int
+testBytesSize = 250000
+
 bm :: Benchmark
 bm =
   bgroup
     "ByteString Encoding to/from integer"
-    [ env (randomBytes 250000) (\bs -> bench "encode bytes to integer" (nf Encoding.encodeByteString bs)),
-      env (Encoding.encodeByteString <$> randomBytes 250000) (\i -> bench "decode bytes from integer" (nf Encoding.decodeByteString i))
+    [ env
+        (randomBytes testBytesSize)
+        (\bs -> bench "encode bytes to integer" (nf Encoding.encodeByteString bs)),
+      env
+        (Encoding.encodeByteString <$> randomBytes testBytesSize)
+        (\i -> bench "decode bytes from integer" (nf (Encoding.decodeByteStringWithDefault (error "failed to decode")) i))
     ]