--- a/src/Juvix/Compiler/Nockma/Encoding/Base.hs
+++ b/src/Juvix/Compiler/Nockma/Encoding/Base.hs
@@ -2,7 +2,6 @@ module Juvix.Compiler.Nockma.Encoding.Base where
 
 import Data.Bit as Bit
 import Data.Bits
-import Data.Vector.Unboxed qualified as U
 import Juvix.Compiler.Nockma.Encoding.Effect.BitReader
 import Juvix.Compiler.Nockma.Encoding.Effect.BitWriter
 import Juvix.Prelude.Base
@@ -12,33 +11,28 @@ import Juvix.Prelude.Base
 writeIntegral :: forall a r. (Integral a, Member BitWriter r) => a -> Sem r ()
 writeIntegral x
   | x < 0 = error "integerToVectorBits: negative integers are not supported in this implementation"
-  | otherwise = unfoldBits (fromIntegral x)
+  | otherwise = unfoldBits 0 (fromIntegral x)
   where
-    unfoldBits :: Integer -> Sem r ()
-    unfoldBits n
-      | n == 0 = return ()
-      | otherwise = writeBit (Bit (testBit n 0)) <> unfoldBits (n `shiftR` 1)
+    len = bitLength x
+
+    unfoldBits :: Int -> Integer -> Sem r ()
+    unfoldBits idx n
+      | idx == len = return ()
+      | otherwise = writeBit (Bit (testBit n idx)) <> unfoldBits (idx + 1) n
+
+-- | Computes the number of bits required to store the argument in binary
+-- NB: 0 is encoded to the empty bit vector (as specified by the Hoon serialization spec), so 0 has bit length 0.
+bitLength :: (Integral a) => a -> Int
+bitLength n
+  | n == 0 = 0
+  | otherwise = fromIntegral (integerLog2 (abs (fromIntegral n))) + 1
 
 integerToVectorBits :: (Integral a) => a -> Bit.Vector Bit
 integerToVectorBits = run . execBitWriter . writeIntegral
 
--- | Computes the number of bits required to store the argument in binary
--- NB: 0 is encoded to the empty bit vector (as specified by the Hoon serialization spec), so 0 has bit length 0.
-bitLength :: forall a. (Integral a) => a -> Int
-bitLength =
-  length
-    . takeWhile (/= 0)
-    . iterate (`shiftR` 1)
-    . toInteger
-
--- | Decode a vector of bits (ordered from least to most significant bits) to an integer
-vectorBitsToInteger :: Bit.Vector Bit -> Integer
-vectorBitsToInteger = U.ifoldl' go 0
-  where
-    go :: Integer -> Int -> Bit -> Integer
-    go acc idx (Bit b)
-      | b = setBit acc idx
-      | otherwise = acc
+-- | Decode a vector of bits (ordered from least to most significant bits) to a ByteString
+vectorBitsToByteString :: Bit.Vector Bit -> ByteString
+vectorBitsToByteString = cloneToByteString
 
 -- | Transform a Natural to an Int, computes Nothing if the Natural does not fit in an Int
 safeNaturalToInt :: Natural -> Maybe Int