@paulcadman 2024-08-28
> Can't we solve this just by eliminating the use of shifts on Integers? They are the culprits here, but just bit-testing should be constant time. I changed two functions in `Nockma.Encoding.Base` to:

Thank you, this is a better approach!

I also looked at the low-level library functions `integerToMutableByteArray` et al. They seem quite complicated to use as they involve doing memory allocations and pointer manipulation.