@paulcadman 2024-08-28
> > Can't we solve this just by eliminating the use of shifts on Integers? They are the culprits here, but just bit-testing should be constant time. I changed two functions in `Nockma.Encoding.Base` to:
> 
> Thank you, this is a better approach!

@lukaszcz  I've reverted to the bitvec implementation using your suggestion for decoding. For me the `juvix eval Tests/Swap.juvix` test is still too slow without the chunking of the bytestring when encoding to an Integer (1m45s without chunking vs 20s with chunking). So I have kept the chunking bytestring -> integer function for encoding.