@paulcadman 2024-08-29
> Or maybe we could just use `Data.Serialize`? The serialization/deserialization implementation for `Integer` should be linear there.

anomaEncode 'serializes' an arbitrary ByteString to an Integer. So I'm not sure we can use Data.Serialize because `serialize : Integer -> ByteString` is not surjective?